- 新增原版熔炉配方适配器。 - 修复 NBT 读取问题。 - 新增 CrT 方法：NovaEngUtils#formatDecimal(double value)。 - 提升 MMCE 版本要求至 R51。

+++ src/main/java/github/kasuminova/novaeng/common/hypernet/NetNodeImpl.java
@@ -29,7 +29,7 @@ public class NetNodeImpl extends NetNode {
         if (isWorking()) {
             if (owner.getTicksExisted() % 10 == 0) {
                 float total = 0;
-                for (final Float value : recipeConsumers.values()) {
+                for (final float value : recipeConsumers.values()) {
                     total += value;
                 }
                 computationPointConsumption = total;
@@ -153,7 +153,7 @@ public class NetNodeImpl extends NetNode {
     @Override
     public void readNBT(final NBTTagCompound customData) {
         super.readNBT(customData);
-        this.computationPointConsumption = customData.getInteger("c");
+        this.computationPointConsumption = customData.getFloat("c");
     }
 
     @Override
